+++ core/src/main/scala/io/kirill/playlistoptimizer/core/optimizer/algorithms/operators/operators.scala
@@ -2,6 +2,6 @@ package io.kirill.playlistoptimizer.core.optimizer.algorithms.operators
 
 package object operators {
 
-  final case class Fitness(value: Double) extends AnyVal
+  final case class Fitness(value: BigDecimal) extends AnyVal
 
 }
